- Added NotFound, Rect and RectInt drawers - GetPriority now returns an int - The not found field is now a special shiroi drawer that can be used for any type

+++ Assets/Shiroi/Cutscenes/Editor/Drawers/TypeDrawer.cs
@@ -11,7 +11,7 @@ namespace Shiroi.Cutscenes.Editor.Drawers {
 
         public abstract void Draw(CutsceneEditor editor, CutscenePlayer player, Cutscene cutscene, Rect rect, int tokenIndex, string name, object value, Type valueType, FieldInfo fieldInfo, Setter setter);
 
-        public virtual byte GetPriority() {
+        public virtual int GetPriority() {
             return 0;
         }
